fix(gojira) use unique tag for -k flag too
-k mounts a path. This path might change branch. Upping it again does
not practically change anything, so even the running gojira with the old
tag will point to a new branch. This means that a gojira upped with a
specific folder manually should always have the same name.

This commit makes this feature isomorphic with what GOJIRA_DETECT_LOCAL
does, which is a good thing, since it makes these both do exactly the
same:

* gojira up -k path/to/kong
* cd path/to/kong ; GOJIRA_DETECT_LOCAL gojira up